+++ src/main/java/org/betonquest/betonquest/compatibility/citizens/CitizensVariable.java
@@ -7,36 +7,78 @@ import org.betonquest.betonquest.Instruction;
 import org.betonquest.betonquest.api.Variable;
 import org.betonquest.betonquest.api.profiles.Profile;
 import org.betonquest.betonquest.exceptions.InstructionParseException;
-import org.bukkit.Location;
+import org.betonquest.betonquest.variables.LocationVariable;
 
+import java.util.Arrays;
 import java.util.Locale;
 
 /**
  * Provides information about a citizen npc.
  * <p>
  * Format:
- * {@code %citizen.<id>.<type>%}
+ * {@code %citizen.<id>.<argument>.<mode>.<precision>%}
  * <p>
- * Types:
- * * name - Return citizen name
- * * full_name - Full Citizen name
- * * location  - Return citizen location. x;y;z;world;yaw;pitch
+ * Arguments:<br>
+ * * name - Return citizen name<br>
+ * * full_name - Full Citizen name<br>
+ * * location - Return citizen location, defaults to ulfLong<br>
+ * Modes: refer to LocationVariable documentation for details.<br>
+ *
+ * @see org.betonquest.betonquest.variables.LocationVariable
  */
-@SuppressWarnings("PMD.CommentRequired")
 public class CitizensVariable extends Variable {
+    /**
+     * The minimum number of required arguments when using a CitizensVariable through an Instruction.
+     */
+    private static final int MINIMUM_INSTRUCTION_ARGUMENTS = 3;
+
+    /**
+     * The integer ID of the NPC.
+     */
     private final int npcId;
 
-    private final TYPE key;
+    /**
+     * The type of information to retrieve for the NPC: name, full_name, or location.
+     */
+    private final ARGUMENT key;
 
+    /**
+     * A wrapper for the location property of the NPC.
+     */
+    private final LocationVariable location;
+
+    /**
+     * Construct a new CitizensVariable that allows for resolution of information about a Citizens NPC.
+     *
+     * @param instruction The Instruction.
+     * @throws InstructionParseException If there was an error parsing the Instruction.
+     */
     public CitizensVariable(final Instruction instruction) throws InstructionParseException {
         super(instruction);
 
-        npcId = instruction.getInt();
-        try {
-            key = TYPE.valueOf(instruction.next().toUpperCase(Locale.ROOT));
-        } catch (final IllegalArgumentException e) {
-            throw new InstructionParseException("Invalid Type: " + instruction.current(), e);
+        final String[] splitInstruction = instruction.getInstruction().split("\\.");
+        if (splitInstruction.length < MINIMUM_INSTRUCTION_ARGUMENTS) {
+            throw new InstructionParseException("Not enough arguments, must have at least " + MINIMUM_INSTRUCTION_ARGUMENTS);
+        } else {
+            try {
+                npcId = Integer.parseInt(splitInstruction[1]);
+                if (npcId < 0) {
+                    throw new InstructionParseException("The specified NPC ID was not a positive or zero integer");
+                }
+            } catch (final NumberFormatException e) {
+                throw new InstructionParseException("The specified NPC ID was not a valid integer", e);
+            }
+
+            final String argument = splitInstruction[2].toUpperCase(Locale.ROOT);
+            try {
+                key = ARGUMENT.valueOf(argument);
+            } catch (final IllegalArgumentException e) {
+                throw new InstructionParseException("Specified CitizenVariable argument was not recognized: '" + argument + "'", e);
+            }
         }
+
+        final String newInstruction = String.join(".", Arrays.copyOfRange(splitInstruction, 2, splitInstruction.length));
+        location = new LocationVariable(new Instruction(instruction.getPackage(), null, newInstruction));
     }
 
     @Override
@@ -47,27 +89,30 @@ public class CitizensVariable extends Variable {
             return "";
         }
 
-        switch (key) {
-            case NAME:
-                return npc.getName();
-            case FULL_NAME:
-                return npc.getFullName();
-            case LOCATION:
-                final Location loc = npc.getStoredLocation();
-                return String.format("%.2f;%.2f;%.2f;%s;%.2f;%.2f",
-                        loc.getX(),
-                        loc.getY(),
-                        loc.getZ(),
-                        loc.getWorld().getName(),
-                        loc.getYaw(),
-                        loc.getPitch());
-        }
-        return "";
+        return switch (key) {
+            case NAME -> npc.getName();
+            case FULL_NAME -> npc.getFullName();
+            case LOCATION -> location.getForLocation(npc.getStoredLocation());
+        };
     }
 
-    private enum TYPE {
+    /**
+     * The type of information to retrieve for the NPC: name, full_name, or location.
+     */
+    private enum ARGUMENT {
+        /**
+         * Retrieve the name of the NPC.
+         */
         NAME,
+
+        /**
+         * Retrieve the full name of the NPC.
+         */
         FULL_NAME,
+
+        /**
+         * Retrieve the location of the NPC.
+         */
         LOCATION
     }